1. Added ability to extend abstractContentContext by allowing listeners to get events from "q" and "Q". this should allow you to implement your own graphic stack extensions, in case you wish to extend the graphic capabilities. (like adding adding fonts outside of the CUsedFonts infrastructure, colour spaces whatnot)
2. Added NO_TIFF and NO_DCT precompiled instructions to allow you to compile/build without tiff images support, or dct filter encoding support. in case you are looking to avoid having to link with Libjpeg or Libtiff

+++ PDFWriter/PDFDate.cpp
@@ -219,8 +219,8 @@ void PDFDate::SetToCurrentTime()
 		else
 		{
 			UTC = timeZoneSecondsDifference > 0 ? eLater:eEarlier;
-			HourFromUTC = labs(timeZoneSecondsDifference) / 3600;
-			MinuteFromUTC = (labs(timeZoneSecondsDifference) - (labs(timeZoneSecondsDifference) / 3600)*3600) / 60;
+			HourFromUTC = (int)(labs(timeZoneSecondsDifference) / 3600);
+			MinuteFromUTC = (int)((labs(timeZoneSecondsDifference) - (labs(timeZoneSecondsDifference) / 3600)*3600) / 60);
 		}
 	}
 	else
